config the deepspeed

+++ train_img_im.py
@@ -148,6 +148,7 @@ logger = utils.get_logger(logpath=os.path.join(
     args.save, 'logs'), filepath=os.path.abspath(__file__))
 logger.info(args)
 
+# device = torch.device('cuda:0' if torch.cuda.is_available() else 'cpu')
 device = torch.device(args.local_rank)
 torch.backends.cudnn.benchmark = True
 
@@ -554,8 +555,7 @@ model = ImplicitFlow(
     classifier=args.classifier,
     chn_dim=im_dim
 )
-
-# model.to(device)
+model = model.half()
 ema = utils.ExponentialMovingAverage(model)
 
 
@@ -745,7 +745,7 @@ def train(epoch, model, trn_loader):
         #   compute z = f(x)
         #   maximize log p(x) = log p(z) - log |det df/dx|
 
-        x = x.to(device)
+        x = x.half().to(device)
 
         beta = beta = min(1, global_itr /
                           args.annealing_iters) if args.annealing_iters > 0 else 1.
@@ -857,7 +857,7 @@ def validate(epoch, model, dat_loader, phase, ema=None):
     start = time.time()
     with torch.no_grad():
         for i, (x, y) in enumerate(tqdm(dat_loader)):
-            x = x.to(device)
+            x = x.half().to(device)
             bpd, logits, _, _ = compute_loss(x, model)
             bpd_meter.update(bpd.item(), x.size(0))
